refactor: move ST_IsClosed from validation to native accessors
