+++ python/python/geodatafusion/geo/_validation.pyi
@@ -1,7 +1,3 @@
-class IsClosed:
-    def __init__(self) -> None: ...
-    def __datafusion_scalar_udf__(self) -> object: ...
-
 class IsValid:
     def __init__(self) -> None: ...
     def __datafusion_scalar_udf__(self) -> object: ...
