+++ python/python/geodatafusion/native/_accessors.pyi
@@ -6,6 +6,10 @@ class EndPoint:
     def __init__(self) -> None: ...
     def __datafusion_scalar_udf__(self) -> object: ...
 
+class IsClosed:
+    def __init__(self) -> None: ...
+    def __datafusion_scalar_udf__(self) -> object: ...
+
 class NDims:
     def __init__(self) -> None: ...
     def __datafusion_scalar_udf__(self) -> object: ...
