+++ python/src/udf/geo/mod.rs
@@ -34,7 +34,6 @@ pub(crate) fn geo(m: &Bound<PyModule>) -> PyResult<()> {
     m.add_class::<relationships::PyWithin>()?;
 
     // validation
-    m.add_class::<validation::PyIsClosed>()?;
     m.add_class::<validation::PyIsValid>()?;
     m.add_class::<validation::PyIsValidReason>()?;
