+++ python/src/udf/geo/validation.rs
@@ -1,7 +1,6 @@
-use geodatafusion::udf::geo::validation::{IsClosed, IsValid, IsValidReason};
+use geodatafusion::udf::geo::validation::{IsValid, IsValidReason};
 
 use crate::impl_udf;
 
-impl_udf!(IsClosed, PyIsClosed, "IsClosed");
 impl_udf!(IsValid, PyIsValid, "IsValid");
 impl_udf!(IsValidReason, PyIsValidReason, "IsValidReason");
