+++ python/src/udf/native/accessors.rs
@@ -1,5 +1,5 @@
 use geodatafusion::udf::native::accessors::{
-    CoordDim, EndPoint, GeometryType, M, NDims, NPoints, NumInteriorRings, ST_GeometryType,
+    CoordDim, EndPoint, IsClosed, GeometryType, M, NDims, NPoints, NumInteriorRings, ST_GeometryType,
     StartPoint, X, Y, Z,
 };
 
@@ -11,6 +11,7 @@ impl_udf!(X, PyX, "X");
 impl_udf!(Y, PyY, "Y");
 impl_udf!(Z, PyZ, "Z");
 impl_udf!(M, PyM, "M");
+impl_udf!(IsClosed, PyIsClosed, "IsClosed");
 impl_udf_coord_type_arg!(EndPoint, PyEndPoint, "EndPoint");
 impl_udf_coord_type_arg!(StartPoint, PyStartPoint, "StartPoint");
 impl_udf!(NPoints, PyNPoints, "NPoints");
