+++ rust/geodatafusion/src/udf/native/accessors/is_closed.rs
