+++ rust/geodatafusion/src/udf/native/accessors/mod.rs
@@ -1,5 +1,6 @@
 mod coord_dim;
 mod geometry_type;
+mod is_closed;
 mod line_string;
 mod npoints;
 mod num_interior_rings;
@@ -7,6 +8,7 @@ mod point;
 
 pub use coord_dim::{CoordDim, NDims};
 pub use geometry_type::{GeometryType, ST_GeometryType};
+pub use is_closed::IsClosed;
 pub use line_string::{EndPoint, StartPoint};
 pub use npoints::NPoints;
 pub use num_interior_rings::NumInteriorRings;
@@ -17,6 +19,7 @@ pub fn register(session_context: &datafusion::prelude::SessionContext) {
     session_context.register_udf(NDims.into());
     session_context.register_udf(GeometryType.into());
     session_context.register_udf(ST_GeometryType.into());
+    session_context.register_udf(IsClosed.into());
     session_context.register_udf(EndPoint::default().into());
     session_context.register_udf(StartPoint::default().into());
     session_context.register_udf(NPoints.into());
